requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
allure-pytest==2.13.3
    # via rest-api-tests (pyproject.toml)
allure-python-commons==2.13.3
    # via allure-pytest
annotated-types==0.6.0
    # via pydantic
attrs==23.1.0
    # via allure-python-commons
autoflake==2.3.1
    # via rest-api-tests (pyproject.toml)
betterproto==2.0.0b5
    # via rest-api-tests (pyproject.toml)
black==24.3.0
    # via rest-api-tests (pyproject.toml)
certifi==2023.5.7
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.1.0
    # via requests
click==8.1.7
    # via black
curlify==2.2.1
    # via rest-api-tests (pyproject.toml)
distconfig3==1.0.1
    # via vyper-config
distlib==0.3.8
    # via virtualenv
docopt==0.6.2
    # via records
et-xmlfile==1.1.0
    # via openpyxl
exceptiongroup==1.1.1
    # via pytest
faker==24.2.0
    # via rest-api-tests (pyproject.toml)
filelock==3.13.1
    # via virtualenv
grpcio==1.54.0
    # via grpcio-tools
grpcio-tools==1.54.0
    # via rest-api-tests (pyproject.toml)
grpclib==0.4.3
    # via betterproto
h2==4.1.0
    # via grpclib
hpack==4.0.0
    # via h2
hyperframe==6.0.1
    # via h2
identify==2.5.35
    # via pre-commit
idna==3.4
    # via requests
inflect==6.0.4
    # via sqlacodegen
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via rest-api-tests (pyproject.toml)
jdcal==1.4.1
    # via openpyxl
multidict==6.0.4
    # via grpclib
mypy-extensions==1.0.0
    # via black
nodeenv==1.8.0
    # via pre-commit
openpyxl==2.4.11
    # via records
packaging==23.1
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.2.0
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via
    #   allure-python-commons
    #   pytest
pre-commit==3.6.2
    # via rest-api-tests (pyproject.toml)
protobuf==4.23.0
    # via grpcio-tools
psycopg2==2.9.9
    # via rest-api-tests (pyproject.toml)
pydantic==2.6.4
    # via
    #   inflect
    #   rest-api-tests (pyproject.toml)
pydantic-core==2.16.3
    # via pydantic
pyflakes==3.2.0
    # via autoflake
pyhamcrest==2.1.0
    # via rest-api-tests (pyproject.toml)
pytelegrambotapi==4.16.1
    # via telebot
pytest==7.3.2
    # via
    #   allure-pytest
    #   rest-api-tests (pyproject.toml)
pytest-telegram-notifier==1.4.0
    # via rest-api-tests (pyproject.toml)
python-dateutil==2.8.2
    # via
    #   betterproto
    #   faker
pyyaml==6.0.1
    # via
    #   pre-commit
    #   vyper-config
records==0.5.3
    # via rest-api-tests (pyproject.toml)
requests==2.31.0
    # via
    #   curlify
    #   pytelegrambotapi
    #   rest-api-tests (pyproject.toml)
    #   telebot
six==1.16.0
    # via python-dateutil
sqlacodegen==2.3.0.post1
    # via rest-api-tests (pyproject.toml)
sqlalchemy==1.4.48
    # via
    #   records
    #   sqlacodegen
structlog==24.1.0
    # via rest-api-tests (pyproject.toml)
tablib==3.4.0
    # via records
telebot==0.0.5
    # via rest-api-tests (pyproject.toml)
toml==0.10.2
    # via vyper-config
tomli==2.0.1
    # via
    #   autoflake
    #   black
    #   pytest
typing-extensions==4.8.0
    # via
    #   black
    #   pydantic
    #   pydantic-core
urllib3==2.0.2
    # via requests
virtualenv==20.25.1
    # via pre-commit
vyper-config==1.2.1
    # via rest-api-tests (pyproject.toml)
watchdog==3.0.0
    # via vyper-config

# The following packages are considered to be unsafe in a requirements file:
# setuptools
